refactor: shared 패키지 유틸리티 함수 분리 및 구조 개선
