+++ packages/shared/src/lib/index.ts
@@ -1,6 +1,3 @@
-export * from './cn';
 export * from './axios';
-export * from './cookies';
-export * from './date';
 export { default as TanStackProvider } from './TanStackProvider';
 export { default as ToastProvider } from './ToastProvider';
