+++ packages/shared/src/ui/Button/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 
 import { Slot } from '@radix-ui/react-slot';
-import { cn } from '@repo/shared/lib/cn';
+import { cn } from '@repo/shared/utils';
 import { type VariantProps, cva } from 'class-variance-authority';
 
 const buttonVariants = cva(
